requirements.txt
streamlit==1.39.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
